+++ iot-e2e-tests/common/src/test/java/tests/integration/com/microsoft/azure/sdk/iot/iothub/telemetry/ReceiveMessagesTests.java
@@ -75,90 +75,4 @@ public class ReceiveMessagesTests extends ReceiveMessagesCommon
     {
         receiveMessage(LARGE_MESSAGE_SIZE_IN_BYTES);
     }
-
-    public void receiveMessage(int messageSize) throws Exception
-    {
-        testInstance.identity.getClient().open(false);
-
-        Message serviceMessage = createCloudToDeviceMessage(messageSize);
-        serviceMessage.setMessageId(UUID.randomUUID().toString());
-
-        com.microsoft.azure.sdk.iot.device.MessageCallback callback = new MessageCallback(serviceMessage);
-
-        if (testInstance.protocol == MQTT || testInstance.protocol == MQTT_WS)
-        {
-            callback = new MessageCallbackMqtt(serviceMessage);
-        }
-
-        Success messageReceived = new Success();
-
-        if (testInstance.identity.getClient() instanceof DeviceClient)
-        {
-            ((DeviceClient) testInstance.identity.getClient()).setMessageCallback(callback, messageReceived);
-        }
-        else if (testInstance.identity.getClient() instanceof ModuleClient)
-        {
-            ((ModuleClient) testInstance.identity.getClient()).setMessageCallback(callback, messageReceived);
-        }
-
-        if (testInstance.identity.getClient() instanceof DeviceClient)
-        {
-            testInstance.messagingClient.send(testInstance.identity.getDeviceId(), serviceMessage);
-        }
-        else if (testInstance.identity.getClient() instanceof ModuleClient)
-        {
-            testInstance.messagingClient.send(testInstance.identity.getDeviceId(), ((TestModuleIdentity) testInstance.identity).getModuleId(), serviceMessage);
-        }
-
-        waitForMessageToBeReceived(messageReceived, testInstance.protocol.toString());
-
-        // flakey feature
-        //waitForFeedbackMessage(serviceMessage.getMessageId());
-
-        Thread.sleep(200);
-        testInstance.identity.getClient().close();
-    }
-
-    private void waitForFeedbackMessage(String expectedMessageId) throws InterruptedException, IOException, IotHubException, TimeoutException
-    {
-        final Success feedbackReceived = new Success();
-
-        Function<FeedbackBatch, AcknowledgementType> feedbackMessageProcessor = feedbackBatch ->
-        {
-            for (FeedbackRecord feedbackRecord : feedbackBatch.getRecords())
-            {
-                if (feedbackRecord.getDeviceId().equals(testInstance.identity.getDeviceId())
-                    && feedbackRecord.getOriginalMessageId().equals(expectedMessageId))
-                {
-                    feedbackReceived.setResult(true);
-                    feedbackReceived.callbackWasFired();
-                }
-            }
-
-            return AcknowledgementType.ABANDON;
-        };
-
-        MessageFeedbackProcessorClientOptions messageFeedbackProcessorClientOptions =
-            MessageFeedbackProcessorClientOptions.builder()
-                .build();
-
-        MessageFeedbackProcessorClient messageFeedbackProcessorClient =
-            new MessageFeedbackProcessorClient(iotHubConnectionString, IotHubServiceClientProtocol.AMQPS_WS, feedbackMessageProcessor, messageFeedbackProcessorClientOptions);
-
-        messageFeedbackProcessorClient.start();
-
-        long startTime = System.currentTimeMillis();
-        while (!feedbackReceived.wasCallbackFired())
-        {
-            Thread.sleep(1000);
-
-            if (System.currentTimeMillis() - startTime > FEEDBACK_TIMEOUT_MILLIS)
-            {
-                fail("Timed out waiting on notification for device " + testInstance.identity.getDeviceId());
-            }
-        }
-
-        messageFeedbackProcessorClient.stop();
-        assertTrue(feedbackReceived.getResult());
-    }
 }
